requirements.txt
numpy==1.16.3
scipy==1.3.0
torch==1.1.0
torchvision==0.3.0
tensorboardX
Pillow
tqdm
matplotlib